requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.2
et-xmlfile==1.0.1
idna==2.10
numpy==1.20.1
openpyxl==3.0.7
pandas==1.2.3
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
urllib3==1.26.4

